This version works.  Changed code so that a library is built and `3dldf' is linked against it. Now "touching" target files to prevent repeated compilation, etc.  This is necessary because of the use of `ctangle':  Sometimes, the target files end up being older than the prerequisites unless the former are "touched".  This occurs when the prerequisites are newer, but the changes don't affect the compiled code.  The situation is more complicated with the Bison and Flex input files.
No longer using "timer" files.  Changed the extension of the generated header files from `h' to `h++'.
Actually, `.h' files are generated by `ctangle' and they are renamed to `h++' if the code has changed
or if no `h++' already exists.  Changed the "include" commands in the CWEB files appropriately.