Now calling `compile_parser.sh' to conditionally compile `parser.c++'.
If `parser.o' isn't up-to-date, `compile_parser.sh' queries the user whether
to recompile `parser.c++' or not.

Previously, `parser.c++' was often recompiled even though the code it contains hadn't changed since
the last compilation.  This was a nuisance, because it takes a long time to compile `parser.c++'.