Removed superfluous blank lines.  Fixed some places where |cerr_mutex| was locked before
a |show| function was called.  The |show| functions also try to lock the mutex, so the program
blocked.  Many of these places are within debugging output.  I tried turning debugging on globally,
but it produced so much output that I had to turn it off again.

Many of these errors with |cerr_mutex| hadn't been discovered earlier because the debugging output
in the relevant functions hadn't been turned on in a long time.