Changed inline comparison functions to non-inline ones.
I need to have pointers to these functions and I don't think it would
be possible to have a pointer to an inline function.  g++ didn't issue
an error, so maybe it is possible.  It could also be possible that the
compiler doesn't inline the functions in this case, but it seems to me
that it would be a fairly complex task to have the compiler do this.

I think it's safer to just not declare them inline.