This version works.  It _does not_ work to fill the spherical biangles with black
and have the declination lines, the names of the plotted stars and the dots for the latter
printed in white on top of them.  There seems to be some problem with the order in which
the MetaPost commands in the MetaPost output file `sample3.mp' are executed.

While testing, the results were puzzling.  Some things worked outside of loops but not inside them.
Sometimes partial lines were visible.  Sometimes lines were covered up in the spherical biangles
where a star was plotted but not in the others.  And so on.

I will try a different approach to getting black spherical biangles with white dots and text.