In `plot_star_on_spherical_biangle':  Added the arguments `boolean suppress_dot'
and `boolean suppress_text_label'.  They can be used for stars that shouldn't be plotted
for some reason, e.g., Polaris, which is at the top of the model where there is no room
for a label.

Since the macro is called, the point on the appropriate spherical biangle for the star
is found (i.e., `R[x]') and most importantly, a call to `\starinfo' for it is written to
`sample_5_star_info_combined.tex', so data for the star will appear in `sample5_a4.dvi',
`sample5_a4.ps' and `sample5_a4.pdf', with a note that it's not plotted.