Finished tabs.  I was going to put a tab at the top and bottom of the spherical biangles,
but I remembered that it would not be easy to do this, because the sides of the apparent triangles
are really curves.  It would therefore take too much work to find points along them.  It would be
much easier to use tape to connect the spherical biangles at the ends.