Added code for writing the file `starinfo_constellations.tex'.  It's included in `clstsph1_a4.tex'.
The other files with calls to the TeX macro `\starinfo' are created, directly or indirectly, by the
a GNU 3DLDF macro in `astronmy.lmc'.  Unlike these files, `starinfo_constellations.tex' is generated by
the program `accsdtbs' ("Access database tables").  This is a better method and I may use it for the
other generated files, too.