Working on rotating labels on |Pictures|.  This works now.  However,
I need to solve the problem of the label positions, i.e., "top", "bot", "lft",
etc.  They are not automatically changed and there's no realistic way of doing
this for rotations that aren't multiples of 45 around one axis.  However, I want
to do this for the ones where it is possible.

If all the labels on a picture are specified such that they lie directly
on the point, rotation works.  Shifting the point can be used as an alternative
to specifying a position.  This would make sense if the user knows that he or
she will be rotating the picture.