Finished rule `command: basic_drawing_command glyph_expression with_clause_draw_list' and
function |Scan_Parse::drawing_command_glyph|.  This version works.

Now including `scanprsf.h++' in `pbsndecl.w'.  This required some corrections in order for the program to
compile and link correctly again.

The next thing to do is to make it possible to access the path vector and the paths it contains of a glyph
interactively.