In |Scan_Parse::show_func|:  BUG FIX:  Now handling |Pictures| specially.  Without the code in the
conditional in the corresponding section, the program either blocks or crashes at the end
of the run or both.

It might be neater to pass the type of the template argument |C *c| as another argument
or to make a specialization of this function for |Pictures|.  However, this solution works.