This version works.  Restored previous state, more or less:  It doesn't correspond
to a particular revision.  I've reverted the code for producing METAFONT output instead of
or in addition to MetaPost output but I've retained the code for the `resolve' command.
I've also fixed a bug connected with showing `current_picture' outside of a figure.

I've removed the code for `verbatim_tex' and `verbatim_metafont' but I plan to restore
the former.  I will restore the latter when and if I restore the code for METAFONT output.