This version seems to work.  I've add the type `superellipse' in the 3DLDF language.  That is,
I've added the necessary tokens, non-terminal symbols, parser rules and functions in order to
be able to use objects of this type in 3DLDF programs.

It was necessary to do this because of the way objects are created and destroyed within the parser:
When a `Path' was created and a `Superellipse' object was assigned to it, the new object was a
`Path' and not a `Superellipse'.  Calling virtual functions on it therefore didn't work.

I haven't defined `superellipse_vector'.  I may do this later.