This version works.  Finished the rule for `directiontime' for the case that the `path' doesn't
need to be transformed.  Working on other rules and functions that require calling MetaPost.  I'll implement
them for the case that the `paths' don't need to be transformed and then go back and program the case that
they do.

Also, I will consider calling |call_metapost| for the ones I wrote before with ad hoc code for reading
the MP output.